# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: AI 철학자 상담 앱 API
  description: 철학자 AI와 대화하는 상담 서비스 API (MVP)
  version: 1.0.0
  contact:
    name: AI Counseling App Team
    email: support@aicounseling.app

servers:
  - url: http://localhost:8080/api
    description: 개발 서버
  - url: https://api.aicounseling.app/api
    description: 운영 서버

tags:
  - name: auth
    description: 인증 관련
  - name: users
    description: 사용자 관련
  - name: counselors
    description: 상담사 관련
  - name: sessions
    description: 세션 관련

paths:
  # ===== 인증 =====
  /auth/login/google:
    post:
      tags: [auth]
      summary: Google OAuth 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: Google OAuth 토큰
      responses:
        200:
          description: 구글 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        401:
          description: 인증 실패

  /auth/login/kakao:
    post:
      tags: [auth]
      summary: Kakao OAuth 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: Kakao OAuth 토큰
      responses:
        200:
          description: 카카오 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        401:
          description: 인증 실패

  /auth/login/naver:
    post:
      tags: [auth]
      summary: Naver OAuth 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: Naver OAuth 토큰
      responses:
        200:
          description: 네이버 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        401:
          description: 인증 실패

  /auth/refresh:
    post:
      tags: [auth]
      summary: 토큰 갱신
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        200:
          description: 토큰 갱신 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  # ===== 사용자 =====
  /users/me:
    get:
      tags: [users]
      summary: 내 정보 조회
      security:
        - bearerAuth: []
      responses:
        200:
          description: 사용자 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'

  /users/nickname:
    patch:
      tags: [users]
      summary: 닉네임 변경
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NicknameUpdateRequest'
      responses:
        200:
          description: 변경 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'

  /users/me:
    delete:
      tags: [users]
      summary: 회원 탈퇴
      security:
        - bearerAuth: []
      responses:
        204:
          description: 탈퇴 완료
        400:
          description: 진행 중인 세션이 있음

  # ===== 상담사 =====
  /counselors:
    get:
      tags: [counselors]
      summary: 상담사 목록 조회
      parameters:
        - name: sort
          in: query
          schema:
            type: string
            enum: [popular, rating, recent]
            default: popular
      responses:
        200:
          description: 상담사 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CounselorListItem'

  /counselors/{id}:
    get:
      tags: [counselors]
      summary: 상담사 상세 정보
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 상담사 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CounselorDetail'
        404:
          description: 상담사를 찾을 수 없음

  /counselors/favorites:
    get:
      tags: [counselors]
      summary: 즐겨찾기 상담사 목록
      security:
        - bearerAuth: []
      responses:
        200:
          description: 즐겨찾기 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CounselorListItem'

  /counselors/{id}/favorite:
    post:
      tags: [counselors]
      summary: 상담사 즐겨찾기 추가
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        201:
          description: 즐겨찾기 추가됨
        409:
          description: 이미 즐겨찾기됨

    delete:
      tags: [counselors]
      summary: 상담사 즐겨찾기 제거
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        204:
          description: 즐겨찾기 제거됨

  # ===== 세션 =====
  /sessions:
    get:
      tags: [sessions]
      summary: 내 상담 세션 목록
      security:
        - bearerAuth: []
      parameters:
        - name: bookmarked
          in: query
          description: 북마크한 세션만 조회
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            description: 페이지 번호 (0-based)
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            description: 페이지 크기
      responses:
        200:
          description: 세션 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/SessionListItem'
                  pageInfo:
                    type: object
                    properties:
                      currentPage:
                        type: integer
                        description: 현재 페이지 (0-based)
                      pageSize:
                        type: integer
                      totalElements:
                        type: integer
                      totalPages:
                        type: integer

    post:
      tags: [sessions]
      summary: 새 상담 세션 시작
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [counselorId]
              properties:
                counselorId:
                  type: integer
      responses:
        201:
          description: 세션 생성됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'

  /sessions/{id}:
    delete:
      tags: [sessions]
      summary: 세션 종료
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 세션 종료됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'

  /sessions/{id}/messages:
    get:
      tags: [sessions]
      summary: 세션 메시지 목록
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        200:
          description: 메시지 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  pageInfo:
                    type: object
                    properties:
                      currentPage:
                        type: integer
                        description: 현재 페이지 (0-based)
                      pageSize:
                        type: integer
                      totalElements:
                        type: integer
                      totalPages:
                        type: integer

    post:
      tags: [sessions]
      summary: 메시지 전송 (AI 응답 포함)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  description: 사용자 메시지
      responses:
        200:
          description: AI 응답 포함
          content:
            application/json:
              schema:
                type: object
                properties:
                  userMessage:
                    type: string
                    description: 사용자 메시지 내용
                  aiMessage:
                    type: string
                    description: AI 응답 내용
                  sessionTitle:
                    type: string
                    nullable: true
                    description: 첫 메시지 후 생성된 제목 (첫 메시지 시에만 포함)


  /sessions/{id}/rate:
    post:
      tags: [sessions]
      summary: 세션 평가 (종료 후에만 가능)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 10
                  description: "별점 (1-10, 별 반개 단위)"
                feedback:
                  type: string
                  description: 피드백 텍스트 (선택)
      responses:
        201:
          description: 평가 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CounselorRating'

  /sessions/{id}/bookmark:
    patch:
      tags: [sessions]
      summary: 세션 북마크 토글
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 북마크 토글됨
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: integer
                  bookmarked:
                    type: boolean

  /sessions/{id}/title:
    patch:
      tags: [sessions]
      summary: 세션 제목 수정
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 100
                  description: 새 세션 제목
      responses:
        200:
          description: 제목 수정됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # ===== 인증 =====
    AuthResponse:
      type: object
      description: OAuth 로그인/토큰 갱신 응답
      properties:
        accessToken:
          type: string
          description: JWT 액세스 토큰
        refreshToken:
          type: string
          description: 리프레시 토큰
        userId:
          type: integer
          description: 사용자 ID
        email:
          type: string
          description: 사용자 이메일
        nickname:
          type: string
          description: 사용자 닉네임

    # ===== 사용자 =====
    UserProfileResponse:
      type: object
      description: 사용자 프로필 정보 (OAuth 전용)
      properties:
        email:
          type: string
          description: 사용자 이메일
        nickname:
          type: string
          description: 사용자 닉네임
        profileImageUrl:
          type: string
          nullable: true
          description: OAuth 프로필 이미지 URL
        authProvider:
          type: string
          enum: [GOOGLE, KAKAO, NAVER]
          description: OAuth 제공자
        memberSince:
          type: string
          format: date
          description: 가입일

    NicknameUpdateRequest:
      type: object
      description: 닉네임 변경 요청
      required: [nickname]
      properties:
        nickname:
          type: string
          minLength: 2
          maxLength: 20
          description: 변경할 닉네임

    # ===== 상담사 =====
    CounselorListItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        title:
          type: string
        avatarUrl:
          type: string
          nullable: true
        averageRating:
          type: integer
          minimum: 0
          maximum: 10
          description: "평균 별점 (1-10, 별 반개 단위)"
        totalSessions:
          type: integer

    CounselorDetail:
      allOf:
        - $ref: '#/components/schemas/CounselorListItem'
        - type: object
          properties:
            description:
              type: string
              description: 상담사 상세 설명
            totalRatings:
              type: integer
              description: 총 평가 개수

    # ===== 세션 =====
    SessionListItem:
      type: object
      properties:
        sessionId:
          type: integer
          description: 세션 ID
        title:
          type: string
          description: 세션 제목
        counselorName:
          type: string
          description: 상담사 이름
        lastMessageAt:
          type: string
          format: date-time
          description: 마지막 메시지 시간
        isBookmarked:
          type: boolean
          description: 북마크 여부

    SessionDetail:
      allOf:
        - $ref: '#/components/schemas/SessionListItem'
        - type: object
          properties:
            closedAt:
              type: string
              format: date-time
              description: 세션 종료 시간 (null이면 진행중)

    # ===== 메시지 =====
    Message:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        senderType:
          type: string
          enum: [USER, AI]
        content:
          type: string
        phase:
          type: string
          enum: [ENGAGEMENT, EXPLORATION, INSIGHT, ACTION, CLOSING]
          description: 이 메시지가 속한 상담 단계
        createdAt:
          type: string
          format: date-time

    # ===== 평가 =====
    CounselorRating:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        counselorId:
          type: integer
        sessionId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 10
          description: "별점 (1-10, 별 반개 단위)"
        feedback:
          type: string
        createdAt:
          type: string
          format: date-time

    # ===== 공통 =====
    RsData:
      type: object
      description: 모든 API 응답의 표준 래퍼
      properties:
        resultCode:
          type: string
          description: 결과 코드 (S-1:성공, F-400:잘못된 요청, F-401:인증 실패, F-404:찾을 수 없음, F-500:서버 오류)
        msg:
          type: string
          description: 응답 메시지
        data:
          type: object
          nullable: true
          description: 실제 응답 데이터

    PagedResponse:
      type: object
      description: 페이지네이션된 응답
      properties:
        content:
          type: array
          description: 실제 데이터 목록
        currentPage:
          type: integer
          description: 현재 페이지 (0-based)
        pageSize:
          type: integer
          description: 페이지 크기
        totalElements:
          type: integer
          description: 전체 요소 수
        totalPages:
          type: integer
          description: 전체 페이지 수
        hasNext:
          type: boolean
          description: 다음 페이지 존재 여부
        hasPrevious:
          type: boolean
          description: 이전 페이지 존재 여부